requirements.txt
django==4.1.7
requests==2.28.2
pillow==10.4.0
mysqlclient==1.4.2
python-dotenv==1.0.0
django-debug-toolbar==3.8.1
